requirements.txt
# Testing
coverage
mock
PyHamcrest==1.6
django-nose==1.0
pinocchio==0.3.1
coveralls>=0.1.1
tox
django-oscar-testsupport==0.3.1

# Development
Werkzeug==0.9.4
django-debug-toolbar==0.11
django-extensions==1.2.5
flake8==0.8
yolk
